+++ src/port_diff.rs
@@ -17,7 +17,7 @@ use std::{
 
 use crate::{
     graph::Graph,
-    port::{BoundPort, BoundaryIndex, Port},
+    port::{BoundPort, BoundaryIndex, BoundarySite, Port},
     subgraph::Subgraph,
 };
 use bimap::BiBTreeMap;
@@ -91,21 +91,6 @@ impl<G: Graph> Ord for PortDiff<G> {
     }
 }
 
-/// A boundary port of a diff
-#[derive(Clone, Serialize, Deserialize, From)]
-#[derive_where(Debug; G: Graph, G::Node: Debug, G::PortLabel: Debug)]
-#[serde(bound(serialize = "G: Serialize, G::Node: Serialize, G::PortLabel: Serialize"))]
-#[serde(bound(
-    deserialize = "G: Deserialize<'de>, G::Node: Deserialize<'de>, G::PortLabel: Deserialize<'de>"
-))]
-pub enum BoundaryPort<G: Graph> {
-    /// A site in the internal graph
-    Site(Site<G::Node, G::PortLabel>),
-    /// A sentinel node, marking a boundary that connects two parent diffs
-    /// without a node in itself.
-    Sentinel(usize),
-}
-
 #[derive(Clone, Serialize, Deserialize)]
 #[serde(bound(serialize = "G: Serialize, G::Node: Serialize, G::PortLabel: Serialize"))]
 #[serde(bound(
@@ -118,7 +103,7 @@ pub struct PortDiffData<G: Graph> {
     ///
     /// Each boundary port of `graph` maps to a port in one of the parents,
     /// reachable by following the `IncomingEdgeIndex`.
-    boundary: Vec<(BoundaryPort<G>, IncomingEdgeIndex)>,
+    boundary: Vec<(BoundarySite<G>, IncomingEdgeIndex)>,
 }
 
 /// The incoming edge at a portdiff, given by its index.
@@ -221,9 +206,9 @@ impl<G: Graph> PortDiff<G> {
         self.data.all_incoming()
     }
 
-    // fn all_parents(&self) -> impl Iterator<Item = Self> + '_ {
-    //     self.data.all_parents().map(|p| p.clone().into()).unique()
-    // }
+    pub fn all_parents(&self) -> impl Iterator<Item = Self> + '_ {
+        self.data.all_parents().map(|p| p.clone().into()).unique()
+    }
 
     /// All outgoing edges.
     fn all_outgoing(&self) -> Vec<OutEdge<G>> {
@@ -281,19 +266,59 @@ impl<G: Graph> PortDiff<G> {
         parent_port.owner.descendants(parent_port.data.opposite())
     }
 
+    /// Resolve a port to a concrete port.
+    ///
+    /// In general, ports may refer to "BoundarySite::Wire"s, which are
+    /// phantom objects (noops) that link two edges together. This function
+    /// resolves such ports to a "concrete" port in a graph by following the
+    /// wire until it finds a real port. As a consequence, this may return 0, 1
+    /// or multiple ports.
+    ///
+    /// If the port is already a concrete port, it is returned as is.
+    pub fn resolve_port(&self, port: Port<G>) -> Vec<Owned<Port<G>, G>> {
+        let boundary = match port {
+            Port::Boundary(index) => index,
+            port @ Port::Bound(..) => {
+                return vec![Owned {
+                    data: port,
+                    owner: self.clone(),
+                }]
+            }
+        };
+        match self.boundary_site(boundary) {
+            BoundarySite::Site(..) => vec![Owned {
+                data: Port::Boundary(boundary),
+                owner: self.clone(),
+            }],
+            &BoundarySite::Wire { id, end } => {
+                let opp_site = BoundarySite::Wire {
+                    id,
+                    end: end.opposite(),
+                };
+                let Some(bd_index) = self
+                    .boundary_iter()
+                    .filter(|&bd| self.boundary_site(bd) == &opp_site)
+                    .at_most_one()
+                    .ok()
+                    .expect("found more than one wire end")
+                else {
+                    return Vec::new();
+                };
+                // Resolve recursively
+                self.opposite_ports(Port::Boundary(bd_index))
+                    .into_iter()
+                    .flat_map(|Owned { data, owner }| owner.resolve_port(data))
+                    .collect()
+            }
+        }
+    }
+
     pub fn boundary_iter(&self) -> impl Iterator<Item = BoundaryIndex> {
         (0..self.boundary.len()).map_into()
     }
 
     /// The boundary port at `boundary`, if it is a site.
-    pub fn boundary_site(&self, boundary: BoundaryIndex) -> Option<&Site<G::Node, G::PortLabel>> {
-        match &self.boundary[usize::from(boundary)].0 {
-            BoundaryPort::Site(site) => Some(site),
-            BoundaryPort::Sentinel(_) => None,
-        }
-    }
-
-    pub fn boundary_port(&self, boundary: BoundaryIndex) -> &BoundaryPort<G> {
+    pub fn boundary_site(&self, boundary: BoundaryIndex) -> &BoundarySite<G> {
         &self.boundary[usize::from(boundary)].0
     }
 
@@ -396,7 +421,7 @@ mod tests {
     use portgraph::{LinkMut, NodeIndex, PortGraph, PortMut, PortOffset};
     use rstest::{fixture, rstest};
 
-    use crate::{port::EdgeEnd, GraphView};
+    use crate::{port::EdgeEnd, PortDiffGraph};
 
     use super::*;
 
@@ -545,7 +570,7 @@ mod tests {
     #[rstest]
     fn serialize_parent_child(parent_child_diffs: [TestPortDiff; 2]) {
         let [_, child] = parent_child_diffs;
-        let graph = GraphView::from_sinks(vec![child]);
+        let graph = PortDiffGraph::from_sinks(vec![child]);
         let serialized = serde_json::to_string_pretty(&graph).unwrap();
         insta::assert_snapshot!(serialized);
     }
@@ -554,7 +579,7 @@ mod tests {
     #[rstest]
     fn serialize_parent_two_children(parent_two_children_diffs: [TestPortDiff; 3]) {
         let [_, child_1, child_2] = parent_two_children_diffs;
-        let graph = GraphView::from_sinks(vec![child_1, child_2]);
+        let graph = PortDiffGraph::from_sinks(vec![child_1, child_2]);
         let serialized = serde_json::to_string_pretty(&graph).unwrap();
         insta::assert_snapshot!(serialized);
     }
@@ -565,7 +590,7 @@ mod tests {
         parent_two_children_overlapping_diffs: [TestPortDiff; 3],
     ) {
         let [_, child_1, child_2] = parent_two_children_overlapping_diffs;
-        let graph = GraphView::from_sinks(vec![child_1, child_2]);
+        let graph = PortDiffGraph::from_sinks(vec![child_1, child_2]);
         let serialized = serde_json::to_string_pretty(&graph).unwrap();
         insta::assert_snapshot!(serialized);
     }