+++ viewer/shared/src/app.rs
@@ -1,6 +1,6 @@
 use crux_core::{render::Render, App};
 use derive_more::{From, Into};
-use portdiff::GraphView;
+use portdiff::PortDiffGraph;
 use portgraph::PortGraph;
 use serde::{Deserialize, Serialize};
 use tket2::static_circ::StaticSizeCircuit;
@@ -38,14 +38,14 @@ impl App for PortDiffViewer {
     fn update(&self, event: Self::Event, model: &mut Self::Model, caps: &Self::Capabilities) {
         match event {
             Event::DeserializeData { data, format } => match format.as_str() {
-                "portgraph" => match serde_json::from_str::<GraphView<PortGraph>>(&data) {
+                "portgraph" => match serde_json::from_str::<PortDiffGraph<PortGraph>>(&data) {
                     Ok(diffs) => model.load(diffs),
                     Err(err) => {
                         caps.log.error(format!("{:?}", err));
                         model.clear()
                     }
                 },
-                "tket" => match serde_json::from_str::<GraphView<StaticSizeCircuit>>(&data) {
+                "tket" => match serde_json::from_str::<PortDiffGraph<StaticSizeCircuit>>(&data) {
                     Ok(diffs) => model.load(diffs),
                     Err(err) => {
                         caps.log.error(format!("{:?}", err));
@@ -79,6 +79,7 @@ impl App for PortDiffViewer {
             graph: "error".to_string(),
             graph_type: "tket",
             hierarchy: vec![],
+            hierarchy_node_labels: vec![],
             selected: Default::default(),
         })
     }
@@ -139,6 +140,7 @@ mod tests {
             graph_type,
             hierarchy,
             selected,
+            ..
         } = view
         else {
             panic!("expected loaded view");
@@ -176,7 +178,7 @@ mod tests {
 
     #[rstest]
     #[case("circ_rewrite.json")]
-    #[case("circ_rewrite_2.json")]
+    #[case("circ_rewrite2.json")]
     fn test_app_load_circuit(#[case] file_name: &str) {
         let app = AppTester::<PortDiffViewer, _>::default();
         let mut model = Model::None;