+++ viewer/shared/src/model.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeSet;
 
 use derive_more::From;
 use petgraph::visit::{EdgeRef, IntoEdgeReferences};
-use portdiff::{self as pd, port_diff::IncompatiblePortDiff, GraphView, NodeId, PortDiff};
+use portdiff::{self as pd, port_diff::IncompatiblePortDiff, NodeId, PortDiff, PortDiffGraph};
 use portgraph::PortGraph;
 use serde::{Deserialize, Serialize};
 use tket2::static_circ::StaticSizeCircuit;
@@ -12,7 +12,7 @@ use crate::{
     DiffId,
 };
 
-type Diffs<G> = GraphView<G>;
+type Diffs<G> = PortDiffGraph<G>;
 type DiffPtr<G> = NodeId<G>;
 
 #[derive(Default, From)]
@@ -71,15 +71,17 @@ impl<G: pd::Graph> LoadedModel<G> {
         let graph_type = graph.graph_type();
         let selected = self.selected_diffs.clone();
         let hierarchy = self.hierarchy().collect();
+        let hierarchy_node_labels = vec![];
         Ok(ViewModel::Loaded {
             graph: graph.to_json(),
             graph_type,
             selected,
             hierarchy,
+            hierarchy_node_labels,
         })
     }
 
-    fn load(all_diffs: GraphView<G>) -> Self {
+    fn load(all_diffs: PortDiffGraph<G>) -> Self {
         let sinks: BTreeSet<DiffPtr<G>> = all_diffs.sinks().map(|d| (&d).into()).collect();
         let mut selected_diffs = BTreeSet::new();
         let mut diff_id_to_ptr = Vec::new();
@@ -122,6 +124,31 @@ impl LoadedModel<StaticSizeCircuit> {
         let circ = PortDiff::extract_graph(diffs).unwrap();
         circ.is_acyclic()
     }
+
+    fn add_hierarchy_node_labels(&self, view: &mut ViewModel) {
+        if let ViewModel::Loaded {
+            hierarchy_node_labels,
+            ..
+        } = view
+        {
+            let root_circ = self
+                .all_diffs
+                .all_nodes()
+                .map(|diff| self.all_diffs.get_diff(diff))
+                .find(|diff| diff.all_parents().next().is_none())
+                .unwrap();
+            let root_cx_count =
+                PortDiff::extract_graph(vec![root_circ]).unwrap().cx_count() as isize;
+            let diff_ptrs = self.diff_id_to_ptr.iter();
+            *hierarchy_node_labels = diff_ptrs
+                .map(|&ptr| self.all_diffs.get_diff(ptr))
+                .map(|diff| {
+                    let g = PortDiff::extract_graph(vec![diff]).unwrap();
+                    (g.cx_count() as isize - root_cx_count).to_string()
+                })
+                .collect();
+        }
+    }
 }
 
 impl Model {
@@ -130,7 +157,11 @@ impl Model {
         match self {
             Model::None => Ok(ViewModel::None),
             Model::Portgraph(model) => model.current_view(),
-            Model::Tket(model) => model.current_view(),
+            Model::Tket(model) => {
+                let mut view = model.current_view()?;
+                model.add_hierarchy_node_labels(&mut view);
+                Ok(view)
+            }
         }
     }
 
@@ -179,6 +210,7 @@ pub enum ViewModel {
         graph: String,
         graph_type: &'static str,
         hierarchy: Vec<HierarchyEdge>,
+        hierarchy_node_labels: Vec<String>,
         selected: BTreeSet<DiffId>,
     },
 }