+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - tbd
+## [0.3.0] - 2024-09-04
+
+### Added
+- Add `PortDiff::resolve_boundary_index` to handle wire boundaries.
+- Add `PortDiffGraph::from_sinks_while` to create a `PortDiffGraph` from a predicate.
+- Add `PortDiffGraph::is_squashable` and `PortDiffGraph::try_squash` to check and squash compatible `PortDiffGraph`s.
+- Add `PortDiff::all_parents` to get all parent `PortDiff`s.
+- Viewer: show CX count delta in hierarchy node labels.
+
+### Changed
+- Redesigned boundary port abstraction. Use `BoundarySite::Wire` to represent a boundary not connected to a site.
+- Rename `GraphView` to `PortDiffGraph` to better reflect its purpose.
+
+### Fixed
+- Crash on squashing several wires with identical IDs.
+- Wires at boundary are squashed properly.
 
 ## [0.2.0] - 2024-09-03
 
@@ -42,7 +57,8 @@ Initial release
 - Serialization and deserialization support
 - Web viewer for PortDiff graphs using Next.js and React
 
-[Unreleased]: https://github.com/lmondada/portdiff/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/lmondada/portdiff/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/lmondada/portdiff/releases/tag/v0.3.0
 [0.2.0]: https://github.com/lmondada/portdiff/releases/tag/v0.2.0
 [0.1.1]: https://github.com/lmondada/portdiff/releases/tag/v0.1.1
 [0.1.0]: https://github.com/lmondada/portdiff/releases/tag/v0.1.0